% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code2hem.R
\name{code2hem}
\alias{code2hem}
\title{Code to HEM name}
\usage{
code2hem(cod)
}
\arguments{
\item{cod}{HEM code.}
}
\value{
Returns the name of the HEM that corresponds to the input code.
}
\description{
This function returns the name of the HEM corresponding to the code of the T2080 manual.
}
\keyword{code}
