% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingTools_data.R
\docType{data}
\name{data_hrm_raw_vignette}
\alias{data_hrm_raw_vignette}
\title{Data generated for HRM merging}
\format{
An object of class dataframe
}
\usage{
data_hrm_raw_vignette
}
\description{
Data from a hardware platform experiment
}
\keyword{datasets}
