% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingTools_data.R
\docType{data}
\name{data_much_raw_vignette}
\alias{data_much_raw_vignette}
\title{Data generated for MUCH merging}
\format{
An object of class dataframe
}
\usage{
data_much_raw_vignette
}
\description{
Data from a hardware platform experiment
}
\keyword{datasets}
