% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mvg_params.R
\name{generate_mvg_params}
\alias{generate_mvg_params}
\title{Generate multivariate Gaussian distribution parameters}
\usage{
generate_mvg_params(splitted_data = NULL, cor_matrix = NULL)
}
\arguments{
\item{splitted_data}{Splitted experimental data.}

\item{cor_matrix}{Correlation matrix}
}
\value{
Returns the parameters of the multivariate Gaussian distributions as well as a list of concatenated HEM readings
}
\description{
\code{generate_mvg_params} takes as input the experimental data along its correlation matrix to estimate the Sigma matrix and the vector of means to compute the multivariate Gaussian distribution
}
\examples{
# merge_hems(data = data,
#            anchor_hem = "PROCESSOR_CYCLES",
#            n_runs = 1000,
#            n_subexp = 3,
#            n_pmcs = 6,
#            processed = FALSE)
}
\keyword{MUCH}
