% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRM.R
\name{merge_hems}
\alias{merge_hems}
\title{HRM merge}
\usage{
merge_hems(
  data = NULL,
  anchor_hem = NULL,
  n_runs = NULL,
  n_subexp = NULL,
  n_pmcs = NULL,
  processed = F
)
}
\arguments{
\item{data}{Sample data.}

\item{anchor_hem}{Reference HEM for merging}

\item{n_runs}{Number of runs on each subexperiment}

\item{n_subexp}{Number of subexperiments}

\item{n_pmcs}{Number of PMCS used on the experiment}

\item{processed}{Is the data already processed in the proper format to merge?}
}
\value{
Returns a merged dataframe
}
\description{
This function merges HEMs coming from separate subexperiments into one single dataset. HRM sorts each subexperiment by the anchor HEM, and the substitutes the anchor HEM by its quantiles.
}
\references{
S. Vilardell, I. Serra, R. Santalla, E. Mezzetti, J. Abella and F. J. Cazorla, "HRM: Merging Hardware Event Monitors for Improved Timing Analysis of Complex MPSoCs," in IEEE Transactions on Computer-Aided Design of Integrated Circuits and Systems, vol. 39, no. 11, pp. 3662-3673, Nov. 2020, <doi:10.1109/TCAD.2020.3013051>.
}
\keyword{HRM}
