% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{process_raw_experiments}
\alias{process_raw_experiments}
\title{Process T2080 data}
\usage{
process_raw_experiments(data, n_pmcs)
}
\arguments{
\item{data}{Raw experimental data from T2080.}

\item{n_pmcs}{Number of PMCs used on the experiment}
}
\value{
Returns the processed raw experimental data.
}
\description{
This functions processes the rawexperimental data into a dataframe by blocks separating each subexperiment.
}
\keyword{HRM,}
\keyword{MUCH}
