% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingTools_data.R
\docType{data}
\name{T2080_code2name}
\alias{T2080_code2name}
\title{T2080 HEM codes to HEM names}
\format{
An object of class \code{data.frame} with 262 rows and 2 columns.
}
\usage{
T2080_code2name
}
\description{
Dataset used to transform the codes from the T2080 manual into readable HEM names.
}
\examples{
data(T2080_code2name)
}
\keyword{datasets}
