% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code2hem.R
\name{code2hem}
\alias{code2hem}
\title{Code to HEM name}
\usage{
code2hem(cod = NULL)
}
\arguments{
\item{cod}{HEM code on the T2080 manual.}
}
\value{
Returns the name of the HEM that corresponds to the input code.
}
\description{
\code{code2hem} returns the name of the HEM corresponding to the code of the T2080 manual.
When gathering experimental data the captured HEMs will be coded with a number.
This functions transforms the number into the reference name in the T2080 manual.
}
\examples{
code <- 1
hem <- code2hem(code)
hem

}
\keyword{HEM}
